+++ src/ciayn/elements/Callable.java
@@ -1,0 +1,11 @@
+package ciayn.elements;
+
+import ciayn.elements.signal.Value;
+
+/**
+ * Created by lukas on 24.01.17.
+ */
+public interface Callable {
+    //Callable callback = null;
+    void callbackActualValue(Value value);
+}
