--- src/ciayn/elements/signal/Callbackable.java
@@ -1,9 +1,0 @@
-package ciayn.elements.signal;
-
-/**
- * Created by lukas on 24.01.17.
- */
-public interface Callbackable {
-    //Callbackable callback = null;
-    void callbackActualValue(Value value);
-}
