+++ src/ciayn/elements/signal/Signal.java
@@ -1,8 +1,10 @@
 package ciayn.elements.signal;
 
 
+import ciayn.elements.Callable;
 import ciayn.elements.Input;
 import ciayn.elements.Output;
+import ciayn.elements.Updatable;
 
 import java.util.List;
 
@@ -11,11 +13,11 @@ import java.util.List;
  */
 
 
-public abstract class Signal implements Updatable,Callbackable  {
+public abstract class Signal implements Updatable,Callable {
     protected Input input;
     protected Output output;
     public abstract List getSignalValues();
-    public abstract void setUpdater(Updater updater);
+    public abstract void setUpdater(Updatable updater);
     public abstract void setActualValue(Number value);
     public abstract void setActualValue(Value value);
     public abstract void setActualValue(Number value, float timeStemp);
