--- src/ciayn/elements/signal/Updater.java
@@ -1,8 +1,0 @@
-package ciayn.elements.signal;
-
-/**
- * Created by lukas on 22.01.17.
- */
-public abstract class Updater implements Updatable {
-    protected Updatable updater = null;
-}
